* 'devel' of https://github.com/arangodb/arangodb: (29 commits)
  make id a string so all of it can be passed on to V8. (#8933)
  Greylist tests/js/client/shell/shell-foxx-paths-spec.js
  fix compile warnings about unused variables/functions
  upgrade velocypack library (#8926)
  Bug fix/cleanup commtasks (#8893)
  Fix example for kShortestPathsGraph
  Small fix of test setUp
  Fix warning about incompatible types in KShortestPathsFinder (#8936)
  Upgrade Swagger UI (#8912)
  Fix samples generation
  Feature/aql skip some (#8771)
  forwardport changes from 3.4 (#8894)
  fix a rare recovery error, add diagnostic output (#8931)
  split headers different, simple cast fixes (#8928)
  fix test
  Suppress replication operations from audit output. (#8896)
  Small speedup of setup in dump suite
  Improve document callback performance (#8874)
  issue 564.1: remove return value and propagate 'force' (#8924)
  fix some zirkus
  ...

+++ lib/Basics/Utf8Helper.cpp
@@ -45,7 +45,7 @@ using namespace icu;
 
 #ifdef _WIN32
 std::wstring arangodb::basics::toWString(std::string const& validUTF8String) {
-  icu::UnicodeString utf16(validUTF8String.c_str(), validUTF8String.size());
+  icu::UnicodeString utf16(validUTF8String.c_str(), static_cast<int32_t>(validUTF8String.size()));
   // // probably required for newer c++ versions
   // using bufferType = std::remove_pointer_t<decltype(utf16.getTerminatedBuffer())>;
   // static_assert(sizeof(std::wchar_t) == sizeof(bufferType), "sizes do not match");
@@ -55,7 +55,7 @@ std::wstring arangodb::basics::toWString(std::string const& validUTF8String) {
 
 std::string arangodb::basics::fromWString(wchar_t const* validUTF16String, std::size_t size) {
   std::string out;
-  icu::UnicodeString ICUString(validUTF16String, size);
+  icu::UnicodeString ICUString(validUTF16String, static_cast<int32_t>(size));
   ICUString.toUTF8String<std::string>(out);
   return out;
 }
